requirements.txt
python-dotenv
Flask
Flask-SQLAlchemy
flask-jwt-extended
flask-admin
werkzeug